+++ src/WebformBuilder.js
@@ -13,7 +13,6 @@ require('./components/builder');
 export default class WebformBuilder extends Webform {
   constructor(element, options) {
     super(element, options);
-    const self = this;
     this.dragContainers = [];
     this.sidebarContainers = [];
     this.updateDraggable = _.debounce(this.refreshDraggable.bind(this), 200);
@@ -36,7 +35,7 @@ export default class WebformBuilder extends Webform {
     this.options.sideBarScroll = _.get(this.options, 'sideBarScroll', true);
     this.options.sideBarScrollOffset = _.get(this.options, 'sideBarScrollOffset', 0);
     this.options.hooks = this.options.hooks || {};
-    this.options.hooks.addComponents = function(components) {
+    this.options.hooks.addComponents = (components) => {
       if (!components || (!components.length && !components.nodrop)) {
         // Return a simple alert so they know they can add something here.
         return [
@@ -56,7 +55,7 @@ export default class WebformBuilder extends Webform {
       }
       return components;
     };
-    this.options.hooks.addComponent = function(container, comp) {
+    this.options.hooks.addComponent = (container, comp) => {
       if (!comp || !comp.component) {
         return container;
       }
@@ -68,12 +67,12 @@ export default class WebformBuilder extends Webform {
         const removeButton = this.ce('div', {
           class: 'btn btn-xxs btn-danger component-settings-button component-settings-button-remove'
         }, this.getIcon('remove'));
-        this.addEventListener(removeButton, 'click', () => self.deleteComponent(comp));
+        this.addEventListener(removeButton, 'click', () => this.deleteComponent(comp));
 
         const editButton = this.ce('div', {
           class: 'btn btn-xxs btn-default component-settings-button component-settings-button-edit'
         }, this.getIcon('cog'));
-        this.addEventListener(editButton, 'click', () => self.editComponent(comp));
+        this.addEventListener(editButton, 'click', () => this.editComponent(comp));
 
         // Add the edit buttons to the component.
         comp.prepend(this.ce('div', {
@@ -82,7 +81,7 @@ export default class WebformBuilder extends Webform {
       }
 
       if (!container.noDrop) {
-        self.addDragContainer(container, this);
+        this.addDragContainer(container, this);
       }
 
       return container;
@@ -748,10 +747,10 @@ export default class WebformBuilder extends Webform {
       this.dragula.destroy();
     }
     this.dragula = dragula(this.sidebarContainers.concat(this.dragContainers), {
-      copy: function(el) {
+      copy(el) {
         return el.classList.contains('drag-copy');
       },
-      accepts: function(el, target) {
+      accepts(el, target) {
         return !target.classList.contains('no-drop');
       }
     }).on('drop', (element, target, source, sibling) => this.onDrop(element, target, source, sibling));